View has been added.

+++ .eslintrc.js
@@ -19,6 +19,6 @@ module.exports = {
     ],
     "rules": {
         "semi": ["error", "always"],
-        "quotes": ["error", "double"]
+        "quotes": ["error", "single"]
     }
 }
